+++ R/stage_parser.R
@@ -3,6 +3,13 @@
 #' Parses results from stage race standings PDFs, i.e. the PDFs with cumulative
 #' results from multiple stages.
 #'
+#' @result A data.frame; specifically a \code[dplyr]{\link{tbl_df}}
+#'
+#' @section Warning: For PDFs that only contain results from 1-2 days of a stage
+#' race, I have found that this function will garble the within stage rank
+#' columns somewhat. In my limited testing, the results are typically repairable
+#' with minimal work.
+#'
 #' @param file character; file path to PDF
 #' @export
 #' @examples